Исправления

+++ Otdel kadrov/Arxiv/frmWorkerRecover.vb
@@ -29,10 +29,10 @@
         txtOsv.Text = SQLreader.GetValue(5)
         txtStagO.Text = ""
         txtSimPolog.Text = SQLreader.GetValue(7)
-        txtData.Text = SQLreader.GetValue(14)
+        txtData.Text = ""
         txtOtdel.Text = SQLreader.GetValue(9)
 
-        PictureBox1.ImageLocation = SQLreader.GetValue(12)
+        PictureBox1.ImageLocation = Application.StartupPath & SQLreader.GetValue(12)
 
         SQLreader.Close()
         db.closeDbConnection()
@@ -51,14 +51,21 @@
     End Sub
 
     Private Sub btnRecover_Click(sender As Object, e As EventArgs) Handles btnRecover.Click
-        PictureBox1.Image.Save(Application.StartupPath & "\images\" & txtId.Text & ".jpg", Imaging.ImageFormat.Jpeg)
+        If txtData.Text <> "" Then
+            PictureBox1.Image.Save(Application.StartupPath & "\images\" & txtId.Text & ".jpg", Imaging.ImageFormat.Jpeg)
 
-        db.createDbConnection(Application.StartupPath & "/db.sqlite")
-        'Виконуємо пакет запітив до БД
+            db.createDbConnection(Application.StartupPath & "/db.sqlite")
+            'Виконуємо пакет запітив до БД
 
-        db.queryDb("INSERT INTO workers(id, fio, adress, phone, dolgn, obraz, stagob, simp, otdel, stag, picture) select id, fio, adress, phone, dolgn, obraz, stagob, simp, otdel, stag, picture from workersf WHERE id='" & txtId.Text & "';" & "
-                            DELETE FROM workersf WHERE id='" & txtId.Text & "'; 
-                            UPDATE workers SET datap='" & txtData.Text & "' WHERE id='" & txtId.Text & "';")
-        db.closeDbConnection()
+            db.queryDb("INSERT INTO workers(id, fio, adress, phone, dolgn, obraz, stagob, simp, otdel, stag, picture) select id, fio, adress, phone, dolgn, obraz, stagob, simp, otdel, stag, picture from workersf WHERE id='" & txtId.Text & "';" & "
+                                DELETE FROM workersf WHERE id='" & txtId.Text & "'; 
+                                UPDATE workers SET datap='" & txtData.Text & "' WHERE id='" & txtId.Text & "';")
+            db.closeDbConnection()
+
+            frmFindWorker.RefreshData()
+            Me.Close()
+        Else
+            MsgBox("Введіть дату прийому!")
+        End If
     End Sub
 End Class
